+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## LATEST
+## v1.6.7
 
 - Fix `zoomToPoints` behavior when points are not initialized [#123](https://github.com/flekschas/regl-scatterplot/issues/123)
 - Add `isDestroyed` and `isPointsDrawn` to `get()`